requirements.txt
# Core LangChain framework (v1.0+)
langchain>=1.0.0
langchain-anthropic>=0.3.0
langchain-openai>=1.0.0

# Model providers (choose one)
openai>=1.0.0
anthropic>=0.18.0

# LangSmith for tracing
langsmith>=0.1.0

# LangGraph Studio (dev server with in-memory runtime)
langgraph-cli[inmem]>=0.4.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
